Crate workos

The official SDK for interacting with the WorkOS API.

Modules

admin_portal

A module for interacting with the WorkOS Admin Portal.

directory_sync

A module for interacting with the WorkOS Directory Sync API.

mfa

A module for interacting with the WorkOS Multi-factor Authentication (MFA) API.

organizations

A module for interacting with organizations within WorkOS.

passwordless

A module for working with passwordless authentication, namely Magic Link.

sso

A module for interacting with the WorkOS Single Sign-On (SSO) API.

webhooks

A module for receiving webhooks from WorkOS.

Structs

ApiKey

An API key to authenticate with the WorkOS API.

ListMetadata

The metadata for a PaginatedList.

PaginatedList

A paginated list of records.

PaginationParams

The parameters used to control pagination for a given paginated endpoint.

RawAttributes

The raw attributes returned from the Identity or Directory Provider.

Timestamp

A UTC timestamp.

Timestamps

The timestamps for an object.

WorkOs

The WorkOS client.

WorkOsBuilder

A builder for a WorkOS client.

Enums

KnownOrUnknown

KnownOrUnknown is a type that respresents either a known value (Known) or an unknown value (Unknown).

PaginationOrder

The order in which records should be returned when paginating.

WorkOsError

A WorkOS SDK error.

Type Aliases

WorkOsResult

A WorkOS SDK result.