workos::resources::user_management

Struct UserManagementApi

pub struct UserManagementApi<'a> { /* private fields */ }

Implementations

impl<'a> UserManagementApi<'a>

pub async fn get_jwks( &self, client_id: &str, params: GetJwksParams, ) -> Result<JwksResponse, Error>

Get JWKS

Returns the JSON Web Key Set (JWKS) containing the public keys used for verifying access tokens.

pub async fn authenticate_with_password( &self, params: AuthenticateWithPasswordParams, ) -> Result<AuthenticateResponse, Error>

Authenticate

Authenticate a user with a specified authentication method.

pub async fn authenticate_with_code( &self, params: AuthenticateWithCodeParams, ) -> Result<AuthenticateResponse, Error>

Authenticate

Authenticate a user with a specified authentication method.

pub async fn authenticate_with_refresh_token( &self, params: AuthenticateWithRefreshTokenParams, ) -> Result<AuthenticateResponse, Error>

Authenticate

Authenticate a user with a specified authentication method.

pub async fn authenticate_with_magic_auth( &self, params: AuthenticateWithMagicAuthParams, ) -> Result<AuthenticateResponse, Error>

Authenticate

Authenticate a user with a specified authentication method.

pub async fn authenticate_with_email_verification( &self, params: AuthenticateWithEmailVerificationParams, ) -> Result<AuthenticateResponse, Error>

Authenticate

Authenticate a user with a specified authentication method.

pub async fn authenticate_with_totp( &self, params: AuthenticateWithTotpParams, ) -> Result<AuthenticateResponse, Error>

Authenticate

Authenticate a user with a specified authentication method.

pub async fn authenticate_with_organization_selection( &self, params: AuthenticateWithOrganizationSelectionParams, ) -> Result<AuthenticateResponse, Error>

Authenticate

Authenticate a user with a specified authentication method.

pub async fn authenticate_with_device_code( &self, params: AuthenticateWithDeviceCodeParams, ) -> Result<AuthenticateResponse, Error>

Authenticate

Authenticate a user with a specified authentication method.

pub async fn get_authorization_url( &self, params: GetAuthorizationUrlParams, ) -> Result<Value, Error>

Get an authorization URL

Generates an OAuth 2.0 authorization URL to authenticate a user with AuthKit or SSO.

pub async fn create_device( &self, params: CreateDeviceParams, ) -> Result<DeviceAuthorizationResponse, Error>

Get device authorization URL

Initiates the CLI Auth flow by requesting a device code and verification URLs. This endpoint implements the OAuth 2.0 Device Authorization Flow (RFC 8628) and is designed for command-line applications or other devices with limited input capabilities.

pub async fn get_logout_url( &self, params: GetLogoutUrlParams, ) -> Result<Value, Error>

Logout

Logout a user from the current session.

pub async fn revoke_session( &self, params: RevokeSessionParams, ) -> Result<Value, Error>

Revoke Session

Revoke a user session.

pub async fn create_cors_origin( &self, params: CreateCorsOriginParams, ) -> Result<CorsOriginResponse, Error>

Create a CORS origin

Creates a new CORS origin for the current environment. CORS origins allow browser-based applications to make requests to the WorkOS API.

pub async fn get_email_verification( &self, id: &str, params: GetEmailVerificationParams, ) -> Result<EmailVerification, Error>

Get an email verification code

Get the details of an existing email verification code that can be used to send an email to a user for verification.

pub async fn reset_password( &self, params: ResetPasswordParams, ) -> Result<PasswordReset, Error>

Create a password reset token

Creates a one-time token that can be used to reset a user’s password.

pub async fn confirm_password_reset( &self, params: ConfirmPasswordResetParams, ) -> Result<ResetPasswordResponse, Error>

Reset the password

Sets a new password using the token query parameter from the link that the user received. Successfully resetting the password will verify a user’s email, if it hasn’t been verified yet.

pub async fn get_password_reset( &self, id: &str, params: GetPasswordResetParams, ) -> Result<PasswordReset, Error>

Get a password reset token

Get the details of an existing password reset token that can be used to reset a user’s password.

pub async fn list_users( &self, params: ListUsersParams, ) -> Result<UserList, Error>

List users

Get a list of all of your existing users matching the criteria specified.

pub async fn create_user(&self, params: CreateUserParams) -> Result<User, Error>

Create a user

Create a new user in the current environment.

pub async fn get_user_by_external_id( &self, external_id: &str, params: GetUserByExternalIdParams, ) -> Result<User, Error>

Get a user by external ID

Get the details of an existing user by an external identifier.

pub async fn get_user( &self, id: &str, params: GetUserParams, ) -> Result<User, Error>

Get a user

Get the details of an existing user.

pub async fn update_user( &self, id: &str, params: UpdateUserParams, ) -> Result<User, Error>

Update a user

Updates properties of a user. The omitted properties will be left unchanged.

pub async fn delete_user( &self, id: &str, params: DeleteUserParams, ) -> Result<Value, Error>

Delete a user

Permanently deletes a user in the current environment. It cannot be undone.

pub async fn confirm_email_change( &self, id: &str, params: ConfirmEmailChangeParams, ) -> Result<EmailChangeConfirmation, Error>

Confirm email change

Confirms an email change using the one-time code received by the user.

pub async fn send_email_change( &self, id: &str, params: SendEmailChangeParams, ) -> Result<EmailChange, Error>

Send email change code

Sends an email that contains a one-time code used to change a user’s email address.

pub async fn verify_email( &self, id: &str, params: VerifyEmailParams, ) -> Result<VerifyEmailResponse, Error>

Verify email

Verifies an email address using the one-time code received by the user.

pub async fn send_verification_email( &self, id: &str, params: SendVerificationEmailParams, ) -> Result<SendVerificationEmailResponse, Error>

Send verification email

Sends an email that contains a one-time code used to verify a user’s email address.

pub async fn get_user_identities( &self, id: &str, params: GetUserIdentitiesParams, ) -> Result<Vec<UserIdentitiesGetItem>, Error>

Get user identities

Get a list of identities associated with the user. A user can have multiple associated identities after going through identity linking. Currently only OAuth identities are supported. More provider types may be added in the future.

pub async fn list_sessions( &self, id: &str, params: ListSessionsParams, ) -> Result<Vec<UserSessionsListItem>, Error>

List sessions

Get a list of all active sessions for a specific user.

pub async fn list_invitations( &self, params: ListInvitationsParams, ) -> Result<Vec<UserInvite>, Error>

List invitations

Get a list of all of invitations matching the criteria specified.

pub async fn send_invitation( &self, params: SendInvitationParams, ) -> Result<UserInvite, Error>

Send an invitation

Sends an invitation email to the recipient.

pub async fn find_invitation_by_token( &self, token: &str, params: FindInvitationByTokenParams, ) -> Result<UserInvite, Error>

Find an invitation by token

Retrieve an existing invitation using the token.

pub async fn get_invitation( &self, id: &str, params: GetInvitationParams, ) -> Result<UserInvite, Error>

Get an invitation

Get the details of an existing invitation.

pub async fn accept_invitation( &self, id: &str, params: AcceptInvitationParams, ) -> Result<Invitation, Error>

Accept an invitation

Accepts an invitation and, if linked to an organization, activates the user’s membership in that organization.

pub async fn resend_invitation( &self, id: &str, params: ResendInvitationParams, ) -> Result<UserInvite, Error>

Resend an invitation

Resends an invitation email to the recipient. The invitation must be in a pending state.

pub async fn revoke_invitation( &self, id: &str, params: RevokeInvitationParams, ) -> Result<Invitation, Error>

Revoke an invitation

Revokes an existing invitation.

pub async fn list_jwt_template( &self, params: ListJWTTemplateParams, ) -> Result<JWTTemplateResponse, Error>

Get JWT template

Get the JWT template for the current environment.

pub async fn update_jwt_template( &self, params: UpdateJWTTemplateParams, ) -> Result<JWTTemplateResponse, Error>

Update JWT template

Update the JWT template for the current environment.

pub async fn create_magic_auth( &self, params: CreateMagicAuthParams, ) -> Result<MagicAuth, Error>

Create a Magic Auth code

Creates a one-time authentication code that can be sent to the user’s email address. The code expires in 10 minutes. To verify the code, authenticate the user with Magic Auth.

pub async fn get_magic_auth( &self, id: &str, params: GetMagicAuthParams, ) -> Result<MagicAuth, Error>

Get Magic Auth code details

Get the details of an existing Magic Auth code that can be used to send an email to a user for authentication.

pub async fn list_organization_memberships( &self, params: ListOrganizationMembershipsParams, ) -> Result<Vec<UserOrganizationMembership>, Error>

List organization memberships

Get a list of all organization memberships matching the criteria specified. At least one of user_id or organization_id must be provided. By default only active memberships are returned. Use the statuses parameter to filter by other statuses.

pub async fn create_organization_membership( &self, params: CreateOrganizationMembershipParams, ) -> Result<OrganizationMembership, Error>

Create an organization membership

Creates a new active organization membership for the given organization and user.

Calling this API with an organization and user that match an inactive organization membership will activate the membership with the specified role(s).

pub async fn get_organization_membership( &self, id: &str, params: GetOrganizationMembershipParams, ) -> Result<UserOrganizationMembership, Error>

Get an organization membership

Get the details of an existing organization membership.

pub async fn update_organization_membership( &self, id: &str, params: UpdateOrganizationMembershipParams, ) -> Result<UserOrganizationMembership, Error>

Update an organization membership

Update the details of an existing organization membership.

pub async fn delete_organization_membership( &self, id: &str, params: DeleteOrganizationMembershipParams, ) -> Result<Value, Error>

Delete an organization membership

Permanently deletes an existing organization membership. It cannot be undone.

pub async fn deactivate_organization_membership( &self, id: &str, params: DeactivateOrganizationMembershipParams, ) -> Result<OrganizationMembership, Error>

Deactivate an organization membership

Deactivates an active organization membership. Emits an organization_membership.updated event upon successful deactivation.

Deactivating an inactive membership is a no-op and does not emit an event.

Deactivating a pending membership returns an error. This membership should be deleted instead.

See the membership management documentation for additional details.

pub async fn reactivate_organization_membership( &self, id: &str, params: ReactivateOrganizationMembershipParams, ) -> Result<UserOrganizationMembership, Error>

Reactivate an organization membership

Reactivates an inactive organization membership, retaining the pre-existing role(s). Emits an organization_membership.updated event upon successful reactivation.

Reactivating an active membership is a no-op and does not emit an event.

Reactivating a pending membership returns an error. The user needs to accept the invitation instead.

See the membership management documentation for additional details.

pub async fn create_redirect_uri( &self, params: CreateRedirectUriParams, ) -> Result<RedirectUri, Error>

Create a redirect URI

Creates a new redirect URI for an environment.

pub async fn list_user_authorized_applications( &self, user_id: &str, params: ListUserAuthorizedApplicationsParams, ) -> Result<AuthorizedConnectApplicationList, Error>

List authorized applications

Get a list of all Connect applications that the user has authorized.

pub async fn delete_user_authorized_application( &self, application_id: &str, user_id: &str, params: DeleteUserAuthorizedApplicationParams, ) -> Result<Value, Error>

Delete an authorized application

Delete an existing Authorized Connect Application.

pub async fn list_user_api_keys( &self, user_id: &str, params: ListUserApiKeysParams, ) -> Result<UserApiKeyList, Error>

List API keys for a user

Get a list of API keys owned by a specific user.

pub async fn create_user_api_key( &self, user_id: &str, params: CreateUserApiKeyParams, ) -> Result<UserApiKeyWithValue, Error>

Create an API key for a user

Create a new API key owned by a user. The user must have an active membership in the specified organization.