Module workos::sso

A module for interacting with the WorkOS Single Sign-On (SSO) API.

WorkOS Docs: SSO Guide

Structs

AccessToken

An access token that may be exchanged for a Profile.

AuthorizationCode

An authorization code that may be exchanged for an SSO profile and access token.

ClientId

A client ID used to initiate SSO.

Connection

WorkOS Docs: Connection

ConnectionId

The ID of a Connection.

DeleteConnectionParams

The parameters for DeleteConnection.

GetAuthorizationUrlParams

The parameters for GetAuthorizationUrl.

GetProfileAndTokenError

An error returned from GetProfileAndToken.

GetProfileAndTokenParams

The parameters for GetProfileAndToken.

GetProfileAndTokenResponse

The response for GetProfileAndToken.

ListConnectionsParams

The parameters for ListConnections.

Profile

WorkOS Docs: Profile

ProfileId

The ID of a Profile.

Sso

Single Sign-On (SSO).

Enums

ConnectionSelector

The selector to use to determine which connection to use for SSO.

ConnectionState

The state of a Connection.

ConnectionType

The type of a Connection.

DeleteConnectionError

An error returned from DeleteConnection.

GetConnectionError

An error returned from GetConnection.

GetProfileError

An error returned from GetProfile.

Provider

An OAuth provider to use for Single Sign-On (SSO).

Traits

DeleteConnection

WorkOS Docs: Delete a Connection

GetAuthorizationUrl

WorkOS Docs: Get Authorization URL

GetConnection

WorkOS Docs: Get a Connection

GetProfile

WorkOS Docs: Get a User Profile

GetProfileAndToken

WorkOS Docs: Get a Profile and Token

ListConnections

WorkOS Docs: List Connections